requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
Django==5.1.1
djangorestframework==3.15.2
dnspython==2.6.1
exceptiongroup==1.2.2
executing==2.1.0
Faker==28.4.1
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.3.1
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pywin32==306
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
